14. Delete unnecesarry file with class
